Move back PyErrorWidget & PyCellViewer into viewer.
Script was needlessly including PyCellViewer.h it seems. Removed it.
This reverts commit 6963a0baa6a19f538e822d4b254b91123dd1dc66.
