+++ hurricane/src/isobar/meson.build
@@ -76,9 +76,7 @@ isobar_sources = files([
   'PyUnitRule.cpp',
   'PyPhysicalRule.cpp',
   'PyTwoLayersPhysicalRule.cpp',
-  'Script.cpp',
-  'PyCellViewer.cpp',
-  'PyErrorWidget.cpp',
+  'Script.cpp'
 ])
 
 isobar = library(
@@ -88,15 +86,5 @@ isobar = library(
   link_with: [hurricane],
   include_directories: hurricane_includes,
   install: true,
-)
-
-py.extension_module(
-  'Hurricane',
-  isobar_sources,
-  dependencies: [qt_deps, py_deps, boost],
-  link_with: [isobar, hurricane],
-  include_directories: hurricane_includes,
-  install: true,
-  subdir: 'coriolis'
 )
