+++ hurricane/src/meson.build
@@ -20,7 +20,7 @@ py.extension_module(
   'Hurricane',
   isobar_sources,
   dependencies: [qt_deps, py_mod_deps, boost],
-  link_with: [isobar, viewer],
+  link_with: [isobar],
   include_directories: hurricane_includes,
   install: true,
   subdir: 'coriolis'
