+++ hurricane/src/viewer/hurricane/viewer/PyErrorWidget.h
