+++ katana/src/PyKatanaEngine.cpp
@@ -16,8 +16,8 @@
 
 #include "hurricane/isobar/PyNet.h"
 #include "hurricane/isobar/PyCell.h"
-#include "hurricane/isobar/PyCellViewer.h"
 #include "hurricane/viewer/ExceptionWidget.h"
+#include "hurricane/viewer/PyCellViewer.h"
 #include "hurricane/Cell.h"
 #include "katana/PyKatanaEngine.h"
 #include "katana/PyKatanaFlags.h"
